Fix DeBERTa-v3-large smoke test: add sentencepiece, load backbone fp32
- pyproject: add sentencepiece==0.2.2 to the torch extra. DeBERTa-v3 ships
  only spm.model (no tokenizer.json), so AutoTokenizer needs the
  sentencepiece backend; without it the tokenizer fails to load.
- models: load the HF backbone with dtype=torch.float32. transformers>=5
  preserves the checkpoint dtype and microsoft/deberta-v3-large stores fp16
  weights, which mismatch the fp32 head and break training.

+++ twitter/models.py
@@ -48,7 +48,13 @@ class TransformerEncoder(nn.Module):
         super().__init__()
 
         self.config = AutoConfig.from_pretrained(name)
-        self.encoder = AutoModel.from_pretrained(name, config=self.config)
+        # transformers>=5 preserves the checkpoint dtype by default. Some
+        # checkpoints (e.g. microsoft/deberta-v3-large) ship fp16 weights, which
+        # then mismatch the fp32 head and break mixed-precision training. Load
+        # the backbone in fp32 and let Lightning's autocast handle precision.
+        self.encoder = AutoModel.from_pretrained(
+            name, config=self.config, dtype=torch.float32
+        )
 
         self.tokenizer = AutoTokenizer.from_pretrained(name, model_max_length=128)
